requirements.txt
Appium-Python-Client==2.8.1
pytest==7.2.2
pytest-html==3.2.0
pytest-rerunfailures==11.1.1
click==8.1.3
pytest-ordering==0.6
# pip install -i https://mirrors.aliyun.com/pypi/simple/ -r requirements.txt